imagenet_inference/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.18
numpy==1.26.4
Pillow==10.4.0
onnxruntime==1.19.2